% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulanrdb-package.R
\docType{package}
\name{ulanrdb}
\alias{ulanrdb}
\alias{ulanrdb-package}
\title{ulanrdb package}
\description{
Ths package is meant to be used with the ulanr package. This package holds a
local version of the ULAN database constructed by ulanr. See the
\url{https://github.com/mdlincoln/ulanr} for more information.
}

